Convert LWRPs to custom resources & fix the specs

+++ recipes/default.rb
@@ -4,7 +4,7 @@
 # Cookbook:: homebrew
 # Recipe:: default
 #
-# Copyright:: 2011-2016, Chef Software, Inc.
+# Copyright:: 2011-2017, Chef Software, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,22 +19,20 @@
 # limitations under the License.
 #
 
-homebrew_go = "#{Chef::Config[:file_cache_path]}/homebrew_go"
+unless homebrew_exists?
+  homebrew_go = "#{Chef::Config[:file_cache_path]}/homebrew_go"
 
-Chef::Log.debug("Homebrew owner is '#{homebrew_owner}'")
+  remote_file homebrew_go do
+    source node['homebrew']['installer']['url']
+    checksum node['homebrew']['installer']['checksum'] unless node['homebrew']['installer']['checksum'].nil?
+    mode '755'
+  end
 
-remote_file homebrew_go do
-  source node['homebrew']['installer']['url']
-  checksum node['homebrew']['installer']['checksum'] unless node['homebrew']['installer']['checksum'].nil?
-  mode '755'
-  not_if { ::File.exist? '/usr/local/bin/brew' }
-end
-
-execute 'install homebrew' do
-  command homebrew_go
-  environment lazy { { 'HOME' => ::Dir.home(homebrew_owner), 'USER' => homebrew_owner } }
-  user homebrew_owner
-  not_if { ::File.exist? '/usr/local/bin/brew' }
+  execute 'install homebrew' do
+    command homebrew_go
+    environment lazy { { 'HOME' => ::Dir.home(homebrew_owner), 'USER' => homebrew_owner } }
+    user homebrew_owner
+  end
 end
 
 execute 'set analytics' do
